+++ kairos-circuit-logic/Cargo.toml
@@ -5,7 +5,7 @@ edition.workspace = true
 license.workspace = true
 
 [features]
-default = ["std"] 
+default = ["std", "arbitrary", "test-logic"] 
 std = []
 serde = ["dep:serde", "kairos-trie/serde"]
 asn1 = ["dep:kairos-tx"]
