+++ kairos-circuit-logic/src/transactions.rs
@@ -129,19 +129,6 @@ mod arbitrary_bounds {
         pub RandomTransaction,
     );
 
-    #[derive(Debug, Default, Clone, Arbitrary)]
-    #[arbitrary(args = AccountsState)]
-    pub struct TestBatch {
-        #[strategy(collection::vec(any_with::<ValidRandomTransaction>(args.clone()), 1..args.shared.max_batch_size))]
-        pub transactions: Vec<ValidRandomTransaction>,
-    }
-
-    #[derive(Debug, Default, Clone, Arbitrary)]
-    #[arbitrary(args = AccountsState)]
-    pub struct TestBatchSequence {
-        pub batches: Vec<Vec<RandomTransaction>>,
-    }
-
     #[derive(Debug, Default)]
     pub struct PublicKeys(pub Vec<Rc<PublicKey>>);
 
@@ -150,8 +137,6 @@ mod arbitrary_bounds {
         shared: Rc<AccountsStateInner>,
     }
     pub struct AccountsStateInner {
-        pub max_batch_size: usize,
-        pub max_batch_count: usize,
         pub keys: PublicKeys,
         pub l1_accounts: RefCell<HashMap<Rc<PublicKey>, u64>>,
         pub l2_accounts: RefCell<HashMap<Rc<PublicKey>, Account>>,
@@ -183,8 +168,6 @@ mod arbitrary_bounds {
         pub fn new() -> Self {
             AccountsState {
                 shared: Rc::new(AccountsStateInner {
-                    max_batch_size: 100,
-                    max_batch_count: 100,
                     keys: PublicKeys::default(),
                     l1_accounts: RefCell::new(HashMap::new()),
                     l2_accounts: RefCell::new(HashMap::new()),
